+++ prography-crawler/src/main/kotlin/org/prography/batch/service/KakaoReviewBatchService.kt
@@ -3,6 +3,7 @@ package org.prography.batch.service
 import org.prography.batch.domain.BatchConstants.BATCH_DELAY
 import org.prography.batch.domain.BatchConstants.BATCH_SIZE
 import org.prography.kakao.review.service.KakaoReviewService
+import org.prography.restaurant.domain.RawRestaurantData
 import org.prography.restaurant.domain.RawRestaurantDataRepository
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
@@ -11,6 +12,7 @@ import org.springframework.data.domain.PageRequest
 import org.springframework.scheduling.annotation.Scheduled
 import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor
 import org.springframework.stereotype.Service
+import java.util.concurrent.CompletableFuture
 
 @Service
 class KakaoReviewBatchService(
@@ -20,6 +22,7 @@ class KakaoReviewBatchService(
     private val executor: ThreadPoolTaskExecutor,
 ) {
     companion object {
+        private const val SLEEP_MS = 5_000L
         private val log: Logger = LoggerFactory.getLogger(KakaoReviewBatchService::class.java)
     }
 
@@ -40,8 +43,16 @@ class KakaoReviewBatchService(
                         BATCH_SIZE,
                     ),
                 )
+            val futures = batch.map(::processRestaurantAsync)
+            futures.awaitAll()
+
+            val failures = futures.count { it.isCompletedExceptionally }
+            log.info(
+                "✅ All review updates completed. successes={} failures={}",
+                batch.size - failures,
+                failures,
+            )
             kakaoReviewService.updateAllReviewsAsync(batch)
-            log.info("Processed $unprocessedCount/${BATCH_SIZE}")
             batch.forEach { it.kakaoReviewProcessed = true }
             rawRestaurantDataRepository.saveAll(batch)
         }
@@ -63,4 +74,21 @@ class KakaoReviewBatchService(
         }
         return false
     }
+
+    private fun processRestaurantAsync(data: RawRestaurantData): CompletableFuture<Void> =
+        CompletableFuture.runAsync({
+            runCatching {
+                Thread.sleep(SLEEP_MS) // TODO: 네트워크 쿨다운이라면 retry/backoff 로 교체 고려
+                kakaoReviewService.saveKakaoReview(data)
+            }.onFailure { ex ->
+                log.error(
+                    "❌ Failed to process review for id={}, kakaoID = {}",
+                    data.id,
+                    data.kakaoPlaceData?.id,
+                    ex,
+                )
+            }
+        }, executor)
 }
+
+private fun <T> List<CompletableFuture<T>>.awaitAll(): Void? = CompletableFuture.allOf(*toTypedArray()).join()
